# making fast AST rewriter faster