… go/tools/asthelpergen/integration/integration_rewriter_test.go
… go/vt/sqlparser/ast_funcs.go
… go/vt/sqlparser/ast_helper.go
… go/vt/sqlparser/ast_rewriting.go
… go/vt/sqlparser/ast_rewriting_test.go
… go/vt/sqlparser/normalizer.go
… go/vt/sqlparser/rewriter_api.go
… go/vt/sqlparser/rewriter_test.go
… go/vt/sqlparser/walker_test.go
… go/vt/vtgate/planbuilder/ddl.go
… go/vt/vtgate/planbuilder/route_planning.go
… go/vt/vtgate/planbuilder/select.go
… go/vt/vtgate/planbuilder/testdata/from_cases.txt
… go/vt/vtgate/semantics/analyzer.go
… go/vt/vttablet/tabletserver/planbuilder/builder.go
… go/vt/wrangler/materializer.go