<!--
  How to contribute: https://vitess.io/docs/contributing/
  Please first make sure there is an open Issue to discuss the feature/fix suggested in this PR.
  If this is a new feature, please mark the Issue as "RFC".
 -->

<!-- if this PR is Work in Progress please create it as a Draft Pull Request -->

## Description
<!-- A few sentences describing the overall goals of the pull request's commits. -->

Removed closures from the ast re-writer and changed re-writer to not return an error.

There is 60%+ improvement on CPU and 15% in memory allocation.

```
name                        old time/op    new time/op    delta
RewriteLargeExpression/1-8     459ns ± 1%     189ns ±10%  -58.87%  (p=0.008 n=5+5)
RewriteLargeExpression/2-8     460ns ± 2%     183ns ±10%  -60.34%  (p=0.008 n=5+5)
RewriteLargeExpression/3-8    8.79µs ± 1%    3.25µs ± 5%  -63.03%  (p=0.008 n=5+5)
RewriteLargeExpression/4-8    3.41µs ± 4%    1.25µs ± 3%  -63.31%  (p=0.008 n=5+5)
RewriteLargeExpression/5-8     400ns ± 2%     171ns ± 1%  -57.26%  (p=0.008 n=5+5)
RewriteLargeExpression/6-8     121µs ± 3%      40µs ± 7%  -67.11%  (p=0.008 n=5+5)

name                        old alloc/op   new alloc/op   delta
RewriteLargeExpression/1-8      128B ± 0%      128B ± 0%     ~     (all equal)
RewriteLargeExpression/2-8      128B ± 0%      128B ± 0%     ~     (all equal)
RewriteLargeExpression/3-8    1.34kB ± 0%    1.26kB ± 0%   -5.39%  (p=0.008 n=5+5)
RewriteLargeExpression/4-8      544B ± 0%      512B ± 0%   -5.88%  (p=0.008 n=5+5)
RewriteLargeExpression/5-8      128B ± 0%      128B ± 0%     ~     (all equal)
RewriteLargeExpression/6-8    13.0kB ± 0%    12.0kB ± 0%   -7.26%  (p=0.008 n=5+5)

name                        old allocs/op  new allocs/op  delta
RewriteLargeExpression/1-8      5.00 ± 0%      5.00 ± 0%     ~     (all equal)
RewriteLargeExpression/2-8      5.00 ± 0%      5.00 ± 0%     ~     (all equal)
RewriteLargeExpression/3-8      68.0 ± 0%      59.0 ± 0%  -13.24%  (p=0.008 n=5+5)
RewriteLargeExpression/4-8      27.0 ± 0%      23.0 ± 0%  -14.81%  (p=0.008 n=5+5)
RewriteLargeExpression/5-8      5.00 ± 0%      5.00 ± 0%     ~     (all equal)
RewriteLargeExpression/6-8       724 ± 0%       607 ± 0%  -16.16%  (p=0.008 n=5+5)

```
## Checklist
- [x] Tests were added or are not required
- [x] Documentation was added or is not required

## Deployment Notes
<!-- Notes regarding deployment of the contained body of work. These should note any db migrations, etc. -->

## Impacted Areas in Vitess
Components that this PR will affect:

- [X]  Query Serving